+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 ### Added
 - `./runtests` test runner with coverage tracking by default
 
+### Fixed
+- Printing of command aliases in help text
+
 ## [1.17.0] - 2016-05-09
 ### Added
 - Pluggable flag-level help text rendering via `cli.DefaultFlagStringFunc`